fix: restore write_lstr16() and fix narrowing conversion breaking CI
- write_lstr16() was wrongly removed in bdfd7961 as "unused" but is
  called 57x in client.cpp's DD RPC methods — broke every CI build.
- session.cpp MgConnect handshake had an int/uint16_t narrowing
  conversion tripping -Werror -Wconversion on clang (macOS job).

+++ src/network/session.cpp
@@ -2664,8 +2664,9 @@ DispatchResult Session::dispatch(const Frame& f) {
             // mgmt-only connection is already tracked — it just never had a
             // user name attached to it until now.
             if (f.payload.size() >= 2) {
-                std::uint16_t ulen = static_cast<std::uint16_t>(f.payload[0]) |
-                    (static_cast<std::uint16_t>(f.payload[1]) << 8);
+                std::uint16_t ulen = static_cast<std::uint16_t>(
+                    static_cast<unsigned>(f.payload[0]) |
+                    (static_cast<unsigned>(f.payload[1]) << 8));
                 if (f.payload.size() >= static_cast<std::size_t>(2 + ulen) && ulen > 0) {
                     std::string mgUser(reinterpret_cast<const char*>(f.payload.data() + 2), ulen);
                     srv_->set_session_user(sid_, mgUser, "");
